ddtrace/tracer: move servicehash to tracer package, improve names and consistency
