+++ ddtrace/tracer/option.go
@@ -538,10 +538,10 @@ func WithPropagator(p Propagator) StartOption {
 	}
 }
 
-// WithServiceNamePropagation enables the hashed name of this service to be horizontally
-// propagated to other services (e.g. over HTTP requests). This enhances the new ingestion control page
+// WithServicePropagation enables the hashed name of this service to be horizontally
+// propagated to other services (e.g. over HTTP requests). This enhances the ingestion control page
 // so that it can show which exact service is the root of ingested spans, and therefore cost.
-func WithServiceNamePropagation(enabled bool) StartOption {
+func WithServicePropagation(enabled bool) StartOption {
 	return func(c *config) {
 		c.propagateServiceName = enabled
 	}
